requirements.txt
scipy==1.4.1
scikit-learn==1.0.2
scikit-image==0.19.2
matplotlib==3.5.2
numpy
py-ecg-detectors==1.0.2
pandas==1.4.2
neurokit2==0.2.0
joblib==1.1.0
statsmodels==0.13.2
tensorflow==2.3.0
keras==2.4.3
xgboost==1.6.1
audiomentations==0.25.1
